+++ CHANGELOG.md
@@ -4,15 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
-## [0.2.100] - 2025-08-17
+## [0.2.110] - 2025-08-17
 ### Changed
 - Updated NuGet packages
 - .NET version updated to 10
 - Updated GitHub Actions
 ### Added
 - Added GitHub Actions for building and publishing the project
+- Added `WebSocketStreamClient` project
+- Added `WebSocketStreamServer` project
 
 ## [0.2.050] - 2025-08-16
 ### Added